Multiplayer UPDATE!
We got multiplayer working through the Dedicated Server. You are now
able to connect and join the server. Which is great news! We now have to
fix a problem with the Mech's Animations. Updating them on the client.
Trying to do that now. Also Special Thanks to Firefly and Incin for
fixing this and getting it to work!